makes goman stfu
Is this it? This is what I got all those ass whoopings for? I had a dream once. It was a dream that little black boys and little black girls would drink from the river of prosperity, freed from the thirst of oppression. But lo and behold, some four decades later, what have I found but a bunch of trifling, shiftless, good-for-nothing niggas? And I know some of you don't want to hear me say that word. It's the ugliest word in the English language, but that's what I see now: niggas. And you don't want to be a nigga, 'cause niggas are living contradictions! Niggas are full of unfulfilled ambitions! Niggas wax and wane; niggas love to complain! Niggas love to hear themselves talk, but hate to explain! Niggas love being another man's judge and jury! Niggas procrastinate until it's time to worry! Niggas love to be late, niggas hate to hurry! Black Entertainment Television is the worst thing I've ever seen in my life! Usher, "Michael Jackson" is not a genre of music! And now I'd like to talk about Soul Plane. I've seen what's around the corner, I've seen what's over the horizon, and I promise you - you niggas won't have nothing to celebrate. And no, I won't get there with you. I'm going to Canada